+++ docs/src/examples/component/maximal/component.js
@@ -1,8 +1,12 @@
 import { defineComponent, getText } from '@semantic-ui/component';
 
 const css = await getText('./component.css');
-const pageCSS = await getText('./component-page.css');
 const template = await getText('./component.html');
+
+// include some page css alongside the component
+const pageCSS = await getText('./component-page.css');
+
+// move a portion of template to a subtemplate for maximalist effect
 import { buttons } from './buttons.js';
 
 /*
@@ -10,7 +14,13 @@ import { buttons } from './buttons.js';
   portions of a component into separate files
   for "enterprisification"
 */
-import { createComponent, onAttributeChanged, onCreated, onDestroyed, onThemeChanged } from './lifecycle.js';
+import {
+  createComponent,
+  onAttributeChanged,
+  onCreated,
+  onDestroyed,
+  onThemeChanged
+} from './lifecycle.js';
 
 import { defaultSettings } from './config.js';