+++ docs/src/examples/component/templates/color-picker/color-picker.js
@@ -3,12 +3,17 @@ import { defineComponent, getText } from '@semantic-ui/component';
 const css = await getText('./component.css');
 const template = await getText('./component.html');
 
-const createComponent = ({ self, signal, dispatchEvent }) => ({
-  colorOptions: ['#ff0000', '#00ff00', '#0000ff', '#ffff00', '#ff00ff', '#00ffff'],
-  selectedColor: signal('#ff0000'),
+const defaultSettings = {
+  colorOptions: ['#ff0000', '#00ff00', '#0000ff', '#ffff00', '#ff00ff', '#00ffff']
+};
 
+const defaultState = {
+  selectedColor: '#ff0000'
+};
+
+const createComponent = ({ state, dispatchEvent }) => ({
   selectColor(color) {
-    self.selectedColor.set(color);
+    state.selectedColor.value = color;
     dispatchEvent('color-selected', { color });
   },
 });
@@ -21,6 +26,8 @@ const events = {
 
 defineComponent({
   tagName: 'color-picker',
+  defaultState,
+  defaultSettings,
   events,
   template,
   css,